+++ packages/lib/src/components/CashAppPay/services/types.ts
@@ -1,3 +1,5 @@
+import { PaymentAmount } from '../../../types';
+
 export interface ICashAppWindowObject {
     pay({ clientId: string }): Promise<ICashAppSDK>;
 }
@@ -7,7 +9,29 @@ export interface ICashAppWindowObject {
  */
 export interface ICashAppSDK {
     render(target: string | HTMLElement, options?: any): Promise<void>;
-    addEventListener(event: string, callback: Function): void;
+    addEventListener(eventType: CashAppPayEvents, callback: Function): void;
+    removeEventListener(eventType: CashAppPayEvents, callback: Function): void;
     customerRequest(customerRequest: any): Promise<void>;
     restart(): Promise<void>;
 }
+
+export enum CashAppPayEvents {
+    CustomerDismissed = 'CUSTOMER_DISMISSED',
+    CustomerRequestApproved = 'CUSTOMER_REQUEST_APPROVED',
+    CustomerRequestDeclined = 'CUSTOMER_REQUEST_DECLINED',
+    CustomerRequestFailed = 'CUSTOMER_REQUEST_FAILED'
+}
+
+export type CashAppServiceConfig = {
+    environment: string;
+    clientId: string;
+    scopeId: string;
+    amount: PaymentAmount;
+    referenceId?: string;
+    button?: {
+        shape?: 'semiround' | 'round';
+        size?: 'medium' | 'small';
+        theme?: 'dark' | 'light';
+        width?: 'static' | 'full';
+    };
+};
